pyproject.toml
[project]
name = "physician-notetaker-emittr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "en-ner-bc5cdr-md",
    "en-ner-bionlp13cg-md",
    "numpy<2",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "scipy>=1.16.3",
    "sentencepiece>=0.2.1",
    "spacy>=3.7.5",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]

[tool.uv.sources]
en-ner-bc5cdr-md = { url = "https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bc5cdr_md-0.5.4.tar.gz" }
en-ner-bionlp13cg-md = { url = "https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bionlp13cg_md-0.5.4.tar.gz" }
